Change files locations
